track-2-backend/track2/requirement.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
transformers
touch
pandas